DDA-83: Use filter functions to filter diorys
