+++ src/react/features/filters/hooks/useFiltersAreActive.js
@@ -1,0 +1,6 @@
+import { useStore } from '../../../store'
+
+export const useFiltersAreActive = () => {
+  const [{ filters }] = useStore((state) => state.filters)
+  return !!Object.values(filters).filter(({ active }) => active).length
+}
